moved ttest and nova methods in different files
